Sem: CGame prepare, play
Methods in CGame to prepare a new game (PvP, PvE) are now completed,
as well as the method play with the main loop of the game. Question
remains, how will the game be paused and saved mid-game.

+++ semestral/src/CPlayer.h
@@ -39,10 +39,10 @@ public:
 	virtual ~CPlayer() = default;
 
 	/// Draws the given amount of cards from drawing pile to hand.
-	void draw_cards( int amount );
+	void draw_cards( size_t amount );
 
 	/// Discards the given amount of cards from hand to discard pile. Lets human player select, which cards he wants to discard.
-	virtual void discard_selection(int amount ) = 0;
+	virtual void discard_selection( int amount ) = 0;
 
 	/// Discards all cards from hand to discard pile. Does not throw any card selection to human player.
 	void discard_all();
